requirements.txt
Flask==0.11.1
requests==2.11.0
xmltodict==0.10.2
python-dotenv==0.5.1
gunicorn==19.6.0
raven==5.24.3
blinker==1.4
